requirements.txt
pandas
numpy
pyzabbix
flask
jaydebeapi
pdfkit
psycopg2-binary
python-dotenv
tkinterweb
plotly
kaleido
scipy
paramiko